Splitting models for recursive complex types into a separate file
This is a prerequisite to sharing the models between E2E and unit tests.
