+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/Build.AspNetCore/Microsoft.Test.E2E.AspNetCore.OData.csproj
@@ -1268,6 +1268,9 @@
     <Compile Include="..\ModelBuilder\PropertyTests.cs">
       <Link>ModelBuilder\PropertyTests.cs</Link>
     </Compile>
+    <Compile Include="..\ModelBuilder\RecursiveComplexTypesModels.cs" >
+      <Link>ModelBuilder\RecursiveComplexTypesModels.cs</Link>
+    </Compile>
     <Compile Include="..\ModelBuilder\RecursiveComplexTypesTests.cs">
       <Link>ModelBuilder\RecursiveComplexTypesTests.cs</Link>
     </Compile>
